Split Boleto and SEPA billing test cases
Committed-By-Agent: codex
